refactor(rspath): split rspath implementations into four distinct files
Improves modularity and maintainability of the codebase.
